+++ src/stores/DocumentStoreContext.jsx
@@ -257,3 +257,17 @@ export function useTakeRating(documentId, takeId) {
 export function useSetTakeRating() {
   return useDocumentStore((ctx) => ctx.setTakeRating);
 }
+
+export function useSetTakeExportedIDBKey() {
+  return useDocumentStore((ctx) => ctx.setTakeExportedIDBKey);
+}
+
+/**
+ * @param {import('./DocumentStore').DocumentId} documentId
+ * @param {import('./DocumentStore').TakeId} takeId
+ */
+export function useTakeExportedIDBKey(documentId, takeId) {
+  return useDocumentStore(
+    (ctx) => getTakeById(ctx, documentId, takeId)?.exportedIDBKey,
+  );
+}